third_party/blink/web_tests/wpt_internal/battery-status/api-defined.html

<!DOCTYPE html>
<html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="/gen/services/device/public/mojom/battery_monitor.mojom.js"></script>
<script src="resources/mock-battery-monitor.js"></script>
<script>
promise_test(async _ => {
  setAndFireMockBatteryInfo(false, 10, 20, 0.5);
  const battery = await navigator.getBattery();

  assert_not_equals(battery, undefined);
  assert_equals(typeof battery.charging, 'boolean');
  assert_equals(typeof battery.chargingTime, 'number');
  assert_equals(typeof battery.dischargingTime, 'number');
  assert_equals(typeof battery.level, 'number');

  testIfBatteryStatusIsUpToDate(battery);

  assert_equals(typeof battery.onchargingchange, 'object');
  assert_equals(typeof battery.onchargingtimechange, 'object');
  assert_equals(typeof battery.ondischargingtimechange, 'object');
  assert_equals(typeof battery.onlevelchange, 'object');
}, 'verify basic getBattery API support');
</script>
</body>
</html>
